Moving code to metadata module
